Refactor TTS into internal/providers/tts
Move ElevenLabs and Kokoro TTS implementations into a new internal/providers/tts package and introduce a shared ttsBase/player to consolidate common playback logic. Deleted the old providers elevenlabs.go and kokoro.go, renamed tts state files and added an Interrupt atomic flag. Updated imports and callers across hooks, plugins (greeting_conversation, speak), and input sensors to reference the new tts package and to support interrupting TTS when user speech is detected. This centralizes TTS behavior, reduces duplication, and adds TTS interrupt handling for inputs.
